decrypted_service/requirements.txt
pika==1.2.0
python-dotenv==0.19.2